update main lab1 -lab6

+++ src/lab/lab1/l2/Main.java
@@ -2,9 +2,9 @@ package lab.lab1.l2;
 
 public class Main {
     public static void main(String[] args) {
-        System.out.println((46+10)*(10/3));
+        System.out.println((46+10)*((double)10/3));
         int x=46;
-        int y =4;
+        int y=4;
         int z = -15;
         System.out.println(x*y*z);
