+++ src/lab/lab1/l3/Main.java
@@ -3,7 +3,7 @@ package lab.lab1.l3;
 public class Main {
     public static void main(String[] args) {
         int number = 10500;
-        int result = (number/10)/10;
+        double result = ((double)number/10)/10;
         System.out.println(result);
 
     }
